fix(test): stop the WebRTC egress probe timing out its own startup
The probe armed its 20s watchdog at script evaluation, so Electron's cold
start was charged to a budget written for the probe. Under CI contention
that startup alone ran 14.5s, and the probe killed itself before doing any
work — reddening unrelated PRs on the package job.

Arm the watchdog after app.whenReady() so it bounds the probe, and move the
probe onto runBrowserRouteEgressElectron, the launcher the h3, TCP and DNS
probes already use, so startup is bounded by a parent that kills the whole
process tree. The probe's own 20s budget is unchanged.
